Refactored method for getting the entry function into interpreter
The get_entry_function was required by both the C test gen stuff (to name the test) and in the interpreter to generate the counter example inputs.
