+++ src/goto-programs/interpreter.cpp
@@ -57,6 +57,38 @@ void interpretert::operator()()
   }
 }
 
+const exprt &interpretert::get_entry_function(const goto_functionst &gf)
+{
+  typedef goto_functionst::function_mapt function_mapt;
+  const function_mapt &fm = gf.function_map;
+
+  const irep_idt goto_start_id(goto_functionst::entry_point());
+
+  const function_mapt::const_iterator start_function = fm.find(goto_start_id);
+
+  // Check we found the start function
+  assert(start_function != fm.end());
+
+  // Iterate over the instructions to find the function call
+  typedef goto_programt::instructionst instructionst;
+  const instructionst &start_instructions = start_function->second.body.instructions;
+
+  typedef goto_programt::instructiont instructiont;
+  typedef goto_programt::instructionst::const_reverse_iterator rev_instruction_itert;
+
+  rev_instruction_itert last_function_call = std::find_if(
+        start_instructions.rbegin(), start_instructions.rend(),
+        [&](const instructiont &instruction ) {
+    return instruction.code.get_statement() == ID_function_call;
+  });
+
+  assert(last_function_call != start_instructions.rend());
+
+  const code_function_callt &func_call=to_code_function_call(last_function_call->code);
+  const exprt &func_expr = func_call.function();
+  return func_expr;
+}
+
 /*******************************************************************\
 
 Function: interpretert::show_state
